Mixer module fix and other things
Now tts gets you a 16kHz 16bit mono WAV and can choose between storing it
in a file or in a variable.
mixer and audiocontroller modules still works with filepaths. A drastic
change may be needed concering next commits

Fixed the print issue with the loop for mixer module. I was retarded :)